user-service/requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
grpcio==1.39.0
grpcio-tools==1.39.0
psycopg2-binary==2.9.1
